patch RtMidi for ARM

+++ extern/rtmidi/RtMidi.cpp
@@ -1805,7 +1805,7 @@ static void *alsaMidiHandler( void *ptr )
           struct timespec y;
           y.tv_nsec = apiData->lastTime.tv_nsec;
           y.tv_sec = apiData->lastTime.tv_sec;
-          if ( x.tv_nsec < y.tv_nsec ) {
+          if ( (long int) x.tv_nsec < y.tv_nsec ) {
               int nsec = (y.tv_nsec - (int)x.tv_nsec) / 1000000000 + 1;
               y.tv_nsec -= 1000000000 * nsec;
               y.tv_sec += nsec;
